refactor: SM-14 컴포넌트 폴더 모듈화
